backend: Introduce basic directory structure
