+++ api/routes/index.py
@@ -1,0 +1,6 @@
+from app import app
+import flask
+
+@app.route("/")
+def index():
+    return flask.send_file("/www/index.html")
